requirements.txt
streamlit
fastapi
uvicorn
python-jose
pdfplumber
pytesseract
pdf2image
chromadb
llama-index
requests
